move octets functions tbd
